policy-agent/requirements.txt
Flask
flask-pymongo
python-dotenv
openai>=1.3.9
openai-agents
PyYaml
python-time
pytest
coverage
mongomock
chromadb
pypdf
sentence-transformers>=2.2.2
tqdm
huggingface-hub>=0.15.1
